refactor(tests): rename fmriprep test files to match what they actually hold
Two problems, one of them introduced by the previous commit.

test_fmriprep_utils.py and test_fmriprep_utils_mock.py were 777 lines of
builders and helpers with no tests in them at all, yet were named test_*.py and
so were collected by pytest only to yield nothing. They are now _support.py and
_mock_fixtures.py: the leading underscore keeps pytest out and tells a reader
there is nothing to run here.

The _mock / _real split also stopped being true once the shared test bodies
landed. Thirteen of the fourteen real-data tests live in the shared file, not in
test_fmriprep_real.py, and mock-driven tests are spread across two files. The
names now state which datasets a file runs against, which is also the question
to ask when deciding where a new test goes:

  test_both_datasets.py   mock and real   13 real + 16 mock
  test_mock_only.py       mock            29, zero carrying real_data
  test_real_only.py       real             1

Each of those claims was checked by collecting with and without -m real_data.

test_real_only.py is named for the dataset rather than for the golden master so
that a later real-only test that is not a golden master still fits. The
distinction still matters and is now written down in README.md: the mock golden
masters are re-derived from the source files by build_expected_bdata_after_exclude
independently of create_bdata_fmriprep, so a mismatch means production disagrees
with an independent calculation; the real one records create_bdata_fmriprep's own
output, so a mismatch means behaviour changed, not that it broke. That is also
why it is not called a ground truth.

The redundant _fmriprep_ infix is gone; the directory already says it. Renames
use git mv, so history follows. All references across the scripts, README and
TEST_COVERAGE.md are updated, and every scripted path was checked to still exist.
Test counts are unchanged: 289 passed / 1 skipped / 14 deselected by default,
13 passed / 1 skipped under -m real_data.
