+++ tests/mri/fmriprep/scripts/mock/step_2_run_test.sh
@@ -1,4 +1,4 @@
 
 PYTHON_BIN="${PYTHON_BIN:-python}"
 
-"${PYTHON_BIN}" -m pytest ./tests/mri/fmriprep/test_fmriprep_mock.py
+"${PYTHON_BIN}" -m pytest ./tests/mri/fmriprep/test_mock_only.py
